refactor: Restructure functions
