+++ src/xiwen/app.py
@@ -1,31 +1,36 @@
-from utils.commands import cmd_demo, cmd_url
-from utils.config import WELCOME_MESSAGE, MENU_OPTIONS
+import random
+from utils.config import DEMO_MESSAGE, DEMO1, DEMO2, MAIN_MENU_OPTIONS, WELCOME_MESSAGE
+from utils.extract import get_hanzi
+from utils.transform import partition_hanzi
+from utils.analysis import analyse
 
 
 def xiwen():
     """
-    Xiwen main menu loop
+    Main menu loop
+    Prompts user for URL to scan
     """
     print(WELCOME_MESSAGE)
-
     while True:
-        # Main menu - get user command
-        print(MENU_OPTIONS)
-        command = input().upper()
+        # Main menu - get URL from user
+        target = input(MAIN_MENU_OPTIONS)
 
-        if command not in ["D", "U", "Q"]:
-            continue  # Invalid command - repeat options
-
-        elif command == "Q":  # Quit command
+        if target.upper() == "Q":  # Quit
             break
 
-        elif command == "D":  # Demo command
-            cmd_demo()
-            continue
+        if target == "":  # Demo simplified or traditional characters
+            target = random.choice([DEMO1, DEMO2])
+            # Explainer
+            print(DEMO_MESSAGE)
 
-        elif command == "U":  # URL command
-            cmd_url()
-            continue
+        # Run URL
+        hanzi_list = get_hanzi(target)
+        if hanzi_list:
+            # Divide into groups (with duplicates)
+            simplified, traditional, outliers = partition_hanzi(hanzi_list)
+
+            if simplified or traditional or outliers:
+                analyse(hanzi_list, simplified, traditional, outliers)
 
 
 if __name__ == "__main__":
